Replace video tag with iframe
The youtube video would not play in my browser
when I ran this locally. I have not used the video
tag before, but was unable to get it working after
about an hour debugging. The iframe works, so I
switched to that so we can get the design changes
live ASAP. We can always switch back to the video
tag if someone gets it working.