Iterate on site with embedded video (#22)
* Iterate on site with embedded video

Committing for Jeanine.
- Fix ugly/broken CSS that Conor did yesterday :)
- Update copy
- Add proper nav bar
- Add CTA to register to vote

* Replace video tag with iframe

The youtube video would not play in my browser
when I ran this locally. I have not used the video
tag before, but was unable to get it working after
about an hour debugging. The iframe works, so I
switched to that so we can get the design changes
live ASAP. We can always switch back to the video
tag if someone gets it working.

* Move CTA above video and update copy

- In the past, voters in PA required an excuse
  to vote by mail. This is the first general election
  in which any registered voter is allowed to vote
  by mail. It is imperative that we educate people
  about that law change.
- Research has shown that telling people how long
  it takes to complete a task will increase the
  chance that they actually undertake that task.
  Therefore, I restored the 2-minute copy and combined
  it with the Wawa reference to get the best of
  both worlds.